# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DreamCraft API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get info about me",
                "operationId": "5c805a4d8b7e4a6c083a8ae945f91952",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-04-27T17:45:19.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-04-27T17:45:19.000000Z"
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "My First Name"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "example": "My Last Name"
                                                    },
                                                    "nickname": {
                                                        "type": "string",
                                                        "example": "My nickname"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "myaddress@gmail.com"
                                                    },
                                                    "email_verified_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-04-27T17:45:19.000000Z"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": "https://2img.net/u/3412/19/22/63/avatars/110530-94.png"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "user"
                                                    },
                                                    "ban": {
                                                        "type": "string",
                                                        "example": "0"
                                                    },
                                                    "reason_ban": {
                                                        "type": "string",
                                                        "example": "null"
                                                    },
                                                    "category_id": {
                                                        "type": "string",
                                                        "example": "null"
                                                    },
                                                    "category_list": {
                                                        "type": "string",
                                                        "example": "History of subscriptions"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "error": "Something wrong. Try later."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user by id",
                "operationId": "6d59d0fdd73e451768018c3539d2afaa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "Integer value",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-04-27T17:45:19.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-04-27T17:45:19.000000Z"
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "My First Name"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "example": "My Last Name"
                                                    },
                                                    "nickname": {
                                                        "type": "string",
                                                        "example": "My nickname"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "myaddress@gmail.com"
                                                    },
                                                    "email_verified_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-04-27T17:45:19.000000Z"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": "https://2img.net/u/3412/19/22/63/avatars/110530-94.png"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "user"
                                                    },
                                                    "ban": {
                                                        "type": "string",
                                                        "example": "0"
                                                    },
                                                    "reason_ban": {
                                                        "type": "string",
                                                        "example": "null"
                                                    },
                                                    "category_id": {
                                                        "type": "string",
                                                        "example": "null"
                                                    },
                                                    "category_list": {
                                                        "type": "string",
                                                        "example": "History of subscriptions"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "message": "Unauthenticated."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "success": false,
                                            "data": "Category not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Result": {
                                        "summary": "Result",
                                        "value": {
                                            "error": "Something wrong. Try later."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}